Renamed assembly to CommonServices
