+++ src/InfuseMediaLibrary/InfuseMediaLibrary.csproj
@@ -18,7 +18,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\LocalFileSystem\LocalFileSystem.csproj" />
+    <ProjectReference Include="..\CommonServices\CommonServices.csproj" />
   </ItemGroup>
-  
+
 </Project>
