+++ src/InfuseMediaLibrary/Services/InfuseMetadataXmlService.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Logging;
 using CSharpFunctionalExtensions;
 using Kurmann.Videoschnitt.InfuseMediaLibrary.Entities;
-using Kurmann.Videoschnitt.LocalFileSystem.Services;
+using Kurmann.Videoschnitt.CommonServices;
 
 namespace Kurmann.Videoschnitt.InfuseMediaLibrary.Services;
