+++ src/InfuseMediaLibrary/Services/MediaIntegratorService.cs
@@ -1,7 +1,7 @@
 using Microsoft.Extensions.Logging;
 using CSharpFunctionalExtensions;
 using System.Text;
-using Kurmann.Videoschnitt.LocalFileSystem.Services;
+using Kurmann.Videoschnitt.CommonServices;
 
 namespace Kurmann.Videoschnitt.InfuseMediaLibrary.Services;
